feat: cut over /duchamp-works to the new gallery UI (pb: pb-695.15.4)
Replace the hardcoded text-card page at /duchamp-works with the
JSON-driven gallery that was built and reviewed at /duchamp-works-preview.
The preview path is gone; users now land on the new UI at the
canonical URL.

- Delete app/(site)/duchamp-works/page.tsx (the 90-line hardcoded
  artwork list with text-only cards).
- Move app/(site)/duchamp-works-preview/page.tsx to
  app/(site)/duchamp-works/page.tsx (the new gallery grid with
  thumbnail tiles, click-to-open detail modal, JSON-backed data).
- Close pb-695.15.4 (cutover) and pb-695.15 (rework umbrella).

Left in place: components/artworks/artwork-card.tsx. It's no longer
used by this page but it is registered as an MDX component in
components/mdx/markdown-renderer.tsx, so removing it would drop a
component authors can reach from content. Flagging in case a
separate audit decides it's truly dead.
